Bring mlp and moe into variations
Refactor for bringing mlp and moe as variations for swapping.
